refactor: move AppHeader and AccountMenu to features layer
