+++ frontend/src/App.tsx
@@ -4,7 +4,7 @@ import { QueryClientProvider } from '@tanstack/react-query';
 import { Toaster } from 'sonner';
 import { getQueryClient } from '@/shared/lib/getQueryClient';
 import { OverlayProvider } from '@/shared/hooks';
-import { AppHeader } from '@/shared/components/AppHeader/AppHeader';
+import { AppHeader } from '@/features/AppHeader/AppHeader';
 import { ProjectSelectPage } from './domains/Projects/page';
 import { PRListPage } from './domains/PRList/page';
 import { PRDetailPage } from './domains/PRDetail/page';
