+++ frontend/src/shared/components/index.ts
@@ -6,7 +6,6 @@ export { Input } from './Input/Input';
 export { Badge } from './Badge/Badge';
 export { Spinner } from './Spinner/Spinner';
 export { Modal } from './Modal/Modal';
-export { AppHeader } from './AppHeader/AppHeader';
 export { Select } from './Select/Select';
 export { Tabs } from './Tabs/Tabs';
 export { Pagination } from './Pagination/Pagination';
